Progress on Python creation of Parameters
The current goal is that ParameterObject.keep_alive determines whether a parameter
should be removed from the list and garbage collected when all Python references are lost.

When .keep_alive == True (default), the parameter will continue to exist and be accessible through .list().

This does mean, however, that the Python user has no control over when the Parameter is actually deleted.
The downside of this is that callbacks for "deleted" parameters could have side effects.

Another solution could've been to have a .remove_from_list() method,
but this opens the door to confusing/undesired behavior when using a parameter that has been removed from the list.