'Properly' expose C symbols.
Turns out the last (and perhaps only) thing that was needed to expose the C symbols,
was to specify "gnu_symbol_visibility: 'default'" in Meson,
which py.extension_module() otherwise sets to 'hidden'.
The journey wasn't entirely pointless however,
as it gave us a good foundation for an __init__.py file,
that we can use to expand PyCSH (itself) with .py files in the future.